xiugai

+++ elead/src/cn/elead/tool/wc/BaseLineUtil.java
@@ -4,7 +4,10 @@ import java.io.Serializable;
 import java.lang.reflect.InvocationTargetException;
 import java.rmi.RemoteException;
 import java.util.ArrayList;
+
+import org.apache.commons.lang.StringUtils;
 import org.apache.log4j.Logger;
+
 import wt.fc.PersistenceHelper;
 import wt.fc.QueryResult;
 import wt.fc.WTObject;
@@ -34,7 +37,7 @@ import wt.vc.baseline.ManagedBaseline;
  * @version
  *
  */
-@SuppressWarnings("serial")
+
 public class BaseLineUtil implements RemoteAccess, Serializable{
 	private static String CLASSNAME = BaseLineUtil.class.getName();
 	private static Logger logger = LogR.getLogger(CLASSNAME);
@@ -46,16 +49,20 @@ public class BaseLineUtil implements RemoteAccess, Serializable{
 	 * @return
 	 * @throws WTException
 	 */
-	@SuppressWarnings({ "deprecation", "unchecked", "rawtypes", "unused" })
+
 	public static ArrayList getAllBaselines(WTPart part) throws WTException {
 		try {
+
+
+
 			if (!RemoteMethodServer.ServerFlag) {
 				return (ArrayList) RemoteMethodServer.getDefault().invoke("getAllBaselines",BaseLineUtil.class.getName(),null,new Class[] { WTPart.class},
 						new Object[] { part });
 			} else {
 				ArrayList baselinelist =new ArrayList();
-				boolean foronce = wt.session.SessionServerHelper.manager.setAccessEnforced(false);
+				boolean foronce = SessionServerHelper.manager.setAccessEnforced(false);
 				try {
+					if(part != null ){
 						QuerySpec qs =new QuerySpec(ManagedBaseline.class);
 						qs.appendWhere(WTContainerHelper.getWhereContainerIs(part.getContainer()));
 						QueryResult qr = PersistenceHelper.manager.find(qs);
@@ -66,6 +73,7 @@ public class BaseLineUtil implements RemoteAccess, Serializable{
 								baselinelist.add(((ManagedBaseline) obj).getName());
 							}
 						}
+					}
 				} catch (Exception e) {
 					logger.error(CLASSNAME+".getAllBaselines:"+e);
 				}finally{
@@ -89,18 +97,17 @@ public class BaseLineUtil implements RemoteAccess, Serializable{
 	 * @return
 	 * @throws WTException
 	 */
-	@SuppressWarnings("deprecation")
 	public static ManagedBaseline getBaseline(String baselineName ,String containerName) throws WTException {
 		try {
 			if (!RemoteMethodServer.ServerFlag) {
 				return (ManagedBaseline) RemoteMethodServer.getDefault().invoke("getBaseline",BaseLineUtil.class.getName(),null,new Class[] { String.class,String.class},
 						new Object[] { baselineName ,containerName});
 			} else {
-				boolean foronce = wt.session.SessionServerHelper.manager.setAccessEnforced(false);
+				boolean foronce = SessionServerHelper.manager.setAccessEnforced(false);
 				ManagedBaseline baseline =null;
 				try {
-					if(!StringUtil.isEmpty(baselineName)&& !StringUtil.isEmpty(containerName)){
-						WTContainer wtContainer=WCUtil.getWTContainerByName(containerName);
+					if(!StringUtils.isEmpty(baselineName)&& !StringUtils.isEmpty(containerName)){
+						WTContainer wtContainer=WCUtil.getWtContainerByName(containerName);
 						QuerySpec qs =new QuerySpec(ManagedBaseline.class);
 						qs.appendWhere(WTContainerHelper.getWhereContainerIs(wtContainer));
 						qs.appendAnd();
@@ -131,7 +138,7 @@ public class BaseLineUtil implements RemoteAccess, Serializable{
 	/**
 	 * 创建基线
 	 * @author zhangxj
-	 * @param part
+	 * @param container
 	 * @return
 	 * @throws WTException
 	 * @throws WTPropertyVetoException
@@ -139,20 +146,15 @@ public class BaseLineUtil implements RemoteAccess, Serializable{
 	public static ManagedBaseline createBaseLine(WTContainer container ,String baseLineName) throws WTException, WTPropertyVetoException{
 		try {
 			if (!RemoteMethodServer.ServerFlag) {
-				return (ManagedBaseline) RemoteMethodServer.getDefault().invoke("getBaseline",BaseLineUtil.class.getName(),null,new Class[] { WTContainer.class,String.class},
+				return (ManagedBaseline) RemoteMethodServer.getDefault().invoke("createBaseLine",BaseLineUtil.class.getName(),null,new Class[] { WTContainer.class,String.class},
 						new Object[] { container ,baseLineName});
 			}else {
-				boolean foronce = wt.session.SessionServerHelper.manager.setAccessEnforced(false);
+				boolean foronce = SessionServerHelper.manager.setAccessEnforced(false);
 				ManagedBaseline baseline = ManagedBaseline.newManagedBaseline();
 				try {
+					if(container != null && !StringUtils.isEmpty(baseLineName)){
 						Folder folder= null;
-						//logger.debug("part=" + part);
-						logger.debug("baseLineName=" + baseLineName);
-						if ( baseLineName == null) {
-							return null;
-						}
 						baseline.setName(baseLineName);
-						//WTContainer container = part.getContainer();
 						//文件夹获取失败，直接创建基线
 						folder = FolderHelper.service.getFolder(BASELINE_LOCATION, WTContainerRef.newWTContainerRef(container));
 						if( folder == null ) {
@@ -162,6 +164,7 @@ public class BaseLineUtil implements RemoteAccess, Serializable{
 						hashSet.add(baseline);
 						FolderHelper.assignLocation(hashSet, folder);
 						PersistenceHelper.manager.save(baseline);
+					}
 				} catch (Exception e) {
 					logger.error(CLASSNAME+".createBaseLine:"+e);
 				}finally{
@@ -185,13 +188,15 @@ public class BaseLineUtil implements RemoteAccess, Serializable{
 	public static void  addObjectForBaseline(WTObject object,Baseline baseline) {
 		try {
 			if (!RemoteMethodServer.ServerFlag) {
-				RemoteMethodServer.getDefault().invoke("getBaseline",BaseLineUtil.class.getName(),null,new Class[] { WTObject.class,Baseline.class},
+				RemoteMethodServer.getDefault().invoke("addObjectForBaseline",BaseLineUtil.class.getName(),null,new Class[] { WTObject.class,Baseline.class},
 						new Object[] { object ,baseline});
 			}else {
-				boolean foronce = wt.session.SessionServerHelper.manager.setAccessEnforced(false);
+				boolean foronce = SessionServerHelper.manager.setAccessEnforced(false);
 				try {
-					Baselineable baselineable=(Baselineable) object;
-					BaselineHelper.service.addToBaseline(baselineable, baseline);
+					if(object != null && baseline != null){
+						Baselineable baselineable=(Baselineable) object;
+						BaselineHelper.service.addToBaseline(baselineable, baseline);
+					}
 				} catch (WTException e) {
 					logger.error(CLASSNAME+".addObjectForBaseline:"+e);
 				}finally{
@@ -204,7 +209,7 @@ public class BaseLineUtil implements RemoteAccess, Serializable{
 			logger.error(e.getMessage(),e);
 		} 
 	}
-		
+	
         /**
          * 移除基线中的对象
          * @param object
@@ -212,55 +217,26 @@ public class BaseLineUtil implements RemoteAccess, Serializable{
          */
 	   public static void removeObjectfromBaseline(WTObject object, Baseline baseline){
 		   try {
-			if (!RemoteMethodServer.ServerFlag) {
-					RemoteMethodServer.getDefault().invoke("getBaseline",BaseLineUtil.class.getName(),null,new Class[] { WTObject.class,Baseline.class},
-							new Object[] { object ,baseline});
-				}else {
-					boolean foronce = wt.session.SessionServerHelper.manager.setAccessEnforced(false);
-					try {
-						Baselineable baselineable = (Baselineable) object;
-						BaselineHelper.service.removeFromBaseline(baselineable, baseline);
-					} catch (WTException e) {
-						logger.error(CLASSNAME+".removeObjectfromBaseline:"+e);
-					}finally{
-						SessionServerHelper.manager.setAccessEnforced(foronce);
+				if (!RemoteMethodServer.ServerFlag) {
+						RemoteMethodServer.getDefault().invoke("removeObjectfromBaseline",BaseLineUtil.class.getName(),null,new Class[] { WTObject.class,Baseline.class },
+								new Object[] { object ,baseline});
+					}else {
+						boolean foronce = SessionServerHelper.manager.setAccessEnforced(false);
+						try {
+							if(object != null && baseline != null){
+								Baselineable baselineable = (Baselineable) object;
+								BaselineHelper.service.removeFromBaseline(baselineable, baseline);
+							}
+						} catch (WTException e) {
+							logger.error(CLASSNAME+".removeObjectfromBaseline:"+e);
+						}finally{
+							SessionServerHelper.manager.setAccessEnforced(foronce);
+						}
 					}
-				}
-		} catch (RemoteException e) {
-			logger.error(e.getMessage(),e);
-		} catch (InvocationTargetException e) {
-			logger.error(e.getMessage(),e);
-		}
+			} catch (RemoteException e) {
+				logger.error(e.getMessage(),e);
+			} catch (InvocationTargetException e) {
+				logger.error(e.getMessage(),e);
+			}
 	   }
-	   
-	public static void main(String[] args) throws RemoteException, InvocationTargetException, WTException{
-		RemoteMethodServer r = RemoteMethodServer.getDefault();
-		r.setUserName("wcadmin");
-		r.setPassword("wcadmin");
-		if (!RemoteMethodServer.ServerFlag) {
-			RemoteMethodServer.getDefault().invoke("test", BaseLineUtil.class.getName(), null,
-					new Class[] {},
-					new Object[] {});
-		}
-	}
-	
-	public static void test() throws WTException, WTPropertyVetoException{
-		WTPart wtPart= PartUtil.getPartByNumber("HQ11120019000");
-//		WTPart wtPart= PartUtil.getPartByNumber("HQ11100114000");
-//		getAllBaselines(wtPart);
-//		System.out.println("getAllBaselines>>>"+getAllBaselines(wtPart));
-//		WTContainer wtContainer=WCUtil.getWTContainerByName("产品信息库");
-//		System.out.println("	getAllBaselines(wtPart);>>>"+	getAllBaselines(wtPart));
-//		getBaseline("测试基线01","A1511");
-//		System.out.println("getBaseline>>>>>>;"+getBaseline("测试基线01","A1511"));
-//		ManagedBaseline baseline =createBaseLine(wtPart,"测试基线06");
-//		System.out.println("baseline>>>>>"+baseline);
-//		BaselineHelper.service.addToBaseline(wtPart, getPartByNumber("HQ11120019000"));
-//		BaselineHelper.service.removeFromBaseline(wtPart, getBaseline("测试基线01","A1511"));
-//		baseline.setName("");
-//		baseline.setNumber("");
-//		WTDocument wtDocument= DocUtil.getDocumentByNumber("CPKF00000062");
-//		addObjectForBaseline(wtDocument,getBaseline("测试基线05","A1511"));
-//		removeObjectfromBaseline(wtDocument,getBaseline("测试基线05","A1511"));
-	}
 }
